Separate dryrun/debug. Fix installed_files.txt.
-D and -D are separate options now. You can use --debug while
installing/uninstalling, and use -d/--dryrun to see what is going
to be installed/uninstalled.

installed_files.txt is always next to ./install.py (before files
are copied to the installation directory).

installed_files.txt is always writable. Even when it's copied
to a system directory.

Extra dirs/files can be specified for removal during `uninstall`.
So if `~/.config/<app_name>` is listed in `cleanup_dirs`, it
will be uninstalled properly.